+++ src/AtomUI.Desktop.Controls/TimePicker/TimePicker.cs
@@ -39,6 +39,7 @@ public class TimePicker : InfoPickerInput
 
     public static readonly StyledProperty<TimeSpan?> SelectedTimeProperty =
         AvaloniaProperty.Register<TimePicker, TimeSpan?>(nameof(SelectedTime),
+            defaultBindingMode: BindingMode.TwoWay,
             enableDataValidation: true);
 
     public static readonly StyledProperty<TimeSpan?> DefaultTimeProperty =